refactored code, tqdm integration
