+++ READMD.md
@@ -1,0 +1,18 @@
+# ArtStyle
+
+Classifying art using deep learning. Work in progress.
+
+## Code structure
+
+### `scripts`
+
+Contains scripts to download and pre-process the WikiArt dataset obtained from
+[Kaggle Painter by Numbers][1].
+
+### `lib`
+
+Python package containing networks, data loaders etc. for use.
+
+See `mnist.ipynb` for illustration on use.
+
+[1]: https://www.kaggle.com/c/painter-by-numbers/data
