+++ lib/model.py
@@ -1,100 +1,22 @@
 """
-Defines a prototypical model module. It contains:
+Defines the `Model` class which encapsulates training, evaluation, and
+serialization of networks. A `Model` is instantiated with a network, loss funcrtion
+(criterion), and optimizer.
 
-* `Net (torch.nn.Module)`: A pytorch neural network definition.
-* `Data (torch.utils.data.DataSet)`: A pytorch dataset to load instances.
+The class provides the following functions:
+
+* `train()`: Train on a provided dataset.
+* `evaluate()`: Measure accuracy on provided dataset.
+* `predict()`: Output scores of network on given minibatch.
+* `save()/load()`: Save network parameters to file.
 """
+import sys
 
 import torch
 import torch.nn as nn
-import torch.nn.functional as F
 from torch.utils.data import DataLoader
-import torchvision.transforms as transforms
-from torchvision.datasets import MNIST
-from sklearn.preprocessing import LabelBinarizer
 import numpy as np
 from tqdm.autonotebook import tqdm, trange
-
-
-class Net(nn.Module):
-
-    def __init__(self):
-        super().__init__()
-
-        # in: N x 1 x 28 x 28
-        self.conv1 = nn.Conv2d(1, 3, 5)
-        # + relu
-        # in: N x 3 x 24 x 24
-        # + pooling (2, 2)
-        # in: N x 3 x 12 x 12
-        self.conv2 = nn.Conv2d(3, 3, 5)
-        # + relu
-        # in: N x 3 x 8 x 8
-        # + pooling (2, 2)
-        # in: N x 3 x 4 x 4
-        # + view N x 48
-        self.fc1 = nn.Linear(48, 10)
-        # in: N x 10
-        # + softmax
-
-
-    def forward(self, x: torch.Tensor):
-        x = self.conv1(x)
-        x = F.relu(x)
-        x = F.max_pool2d(x, (2, 2))
-
-        x = self.conv2(x)
-        x = F.relu(x)
-        x = F.max_pool2d(x, (2, 2))
-
-        x = x.view(-1, 48)
-        x = self.fc1(x)
-        # x = F.softmax(x, dim=1)
-        return x
-
-
-
-class Data(MNIST):
-    """
-    The `torchvision.datasets.MNIST` class but with transforms predefined so
-    the images are loaded as `1 x 28 x 28` tensors and labels as one-hot coded
-    10-element tensors.
-
-    Args:
-
-    * root (str): The root directory of the image files.
-    * train (bool): Whether to load training or testing data.
-    * download (bool): Whether to download to root if files not present.
-    * `binarize (bool)`: Whether to one-hot code integer labels. False (default)
-    for cross entropy loss.
-    """
-
-    # The following two transform functions are used by torch.utils.data.DataSet
-    # to read MNIST data from a binary file and properly transform into image
-    # tensors and one-hot coded labels for use by the network.
-
-    # Transforms PIL.Image into a torch.Tensor
-    # Any other transforms can be added to the list
-    transform = transforms.Compose([
-        transforms.ToTensor(),
-    ])
-
-
-    # Transforms a single label using one-hot coding into a binary array
-    # pylint: disable=E0602,E1101,E1102
-    # (^ disables incorrect error warning by pylint)
-    binarizer = LabelBinarizer().fit(np.arange(10))
-    target_transform = lambda x: torch.tensor(x, dtype=torch.long)
-    target_transform_b = lambda x: torch.tensor(binarizer.transform(x), dtype=torch.long)
-
-
-    def __init__(self, root: str, train: bool=True, download: bool=True,
-        binarize: bool=False):
-
-        super().__init__(download=download, train=train, root=root,
-            transform=self.__class__.transform,
-            target_transform=self.__class__.target_transform if not binarize \
-                            else self.__class__.target_transform_b)
 
 
 
@@ -147,21 +69,24 @@ class Model:
         trainloader = DataLoader(dataset, batch_size=batch_size)
         self.net.to(self.device)
 
-        for epoch in trange(epochs):
-            tqdm.write('Epoch # {}'.format(epoch))
+        for epoch in trange(epochs, desc='Epochs'):
             running_loss = 0.
-            for i, (batchX, batchY) in enumerate(tqdm(trainloader), 1):
+            for i, (batchX, batchY) in enumerate(
+                    tqdm(trainloader, desc='Epoch {}'.format(epoch+1), leave=False), 1):
+
                 batchX, batchY = batchX.to(self.device), batchY.to(self.device)
-                
+
                 self.optimizer.zero_grad()           # clear gradients from prev. step
                 predY = self.net(batchX)             # get predicted labels
                 loss = self.criterion(predY, batchY) # compute loss
                 loss.backward()                      # backpropagate - populate error grad.
                 self.optimizer.step()                # update weights based on gradients
-                
+
                 running_loss += loss.item()
                 if i % verbosity == 0:
-                    tqdm.write('Min-batch # {0:4d}\t Loss: {1:3.3f}'.format(i, running_loss / verbosity))
+                    avg_loss = running_loss / verbosity
+                    tqdm.write('Mini-batch # {0:4d}\t Loss: {1:3.3f}'.format(i,
+                    avg_loss), file=sys.stderr)
                     running_loss = 0.
         # restore net's training mode
         self.net.train(curr_mode)
@@ -204,7 +129,7 @@ class Model:
         # pylint: disable=E1101
         total = 0
         with torch.no_grad():
-            for batchX, batchY in tqdm(testloader):
+            for batchX, batchY in tqdm(testloader, leave=False):
                 # convert to GPU tensor
                 batchX = batchX.to(self.device)
                 # convert to numpy array of batch_size x labels
@@ -223,6 +148,9 @@ class Model:
 
 
     def save(self, path: str):
+        """
+        Saves network weights and hyperparameters in a pickle file.
+        """
         net_state = self.net.state_dict()
         if self.optimizer is not None:
             opt_state = self.optimizer.state_dict()
@@ -236,6 +164,11 @@ class Model:
 
 
     def load(self, path: str):
+        """
+        Loads network weights and hyperparameters from a pickle file. The model
+        has to be instantiated with the same `net`, `criterion`, and `optimizer`
+        arguments.
+        """
         net_state, opt_state, cri_state = torch.load(path, map_location='cpu')
         self.net.load_state_dict(net_state)
         if self.optimizer is not None:
