+++ lib/net/__init__.py
@@ -1,0 +1,3 @@
+"""
+Defines network architectures (`torch.nn.Module`) to be used in pipelines.
+"""
