--- pipelines/__init__.py
@@ -1,3 +1,0 @@
-from . import mnist
-from . import cropped
-from . import resized
